Fix Enter key event for interactive Expand Abbr.
Because Emmet's Interactive Expand Abbreviations panel shows up when the `Ctrl + Shift + Alt + Enter` (default value) key are pressed (down), when the `Enter` key is up, the panel suddenly closed. So there is no time for typing anything into.

This is a hack for preventing the panel from being closed.